requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
cerebras-cloud-sdk==1.21.0
authlib==1.5.0
httpx==0.28.1
itsdangerous==2.2.0
